Schema (#3)
* .gitignore fix

* moving from wip repo

* Delete Software Requirements Document.md

duplicate

* initialize

* Added BaseX to .gitignore

This is something that I am trying to work on for the
database. This is a pretty interesting thing that could be
helpful if I can figure out how to use it properly. This
will be restricted to the schema branch for the time being
as it involves the xml files.

I also added message.txt to the .gitignore file so I can add
longer messages to my commits.

* Added BaseX to .gitignore

This is something that I am trying to work on for the
database. This is a pretty interesting thing that could be
helpful if I can figure out how to use it properly. This
will be restricted to the schema branch for the time being
as it involves the xml files.

I also added message.txt to the .gitignore file so I can add
longer messages to my commits.

* fix .gitignore

removed schema related ignores

* adding files

* .gitignore fix

* Fix

* Delete test.patch

* Delete .gitignore

* Revert ".gitignore fix"

This reverts commit 5895d435b403b16ef20a50ac19c3698ac32ce9fc.